+++ NEWS.md
@@ -3,7 +3,8 @@
   * Improve title checking
   * Move check_template to create_final
   * Move check_lint to create_final
-  * Try to detect lanaguage
+  * Try to detect programming lanaguage
+  * 
 
 # 0.5.0
   * Extract latex tags using xparse from latex