+++ NEWS.md
@@ -6,6 +6,7 @@
   * Try to detect programming lanaguage
   * Ensure library calls use quotes
   * Check margin notes & captions finish with a full stop
+  * Add the course package version to the final page
 
 # 0.5.0
   * Extract latex tags using xparse from latex